requirements.txt
streamlit
sentence-transformers
pandas
python-docx
python-pptx
PyMuPDF
faiss-cpu
transformers
numpy
langchain_groq
langchain